Change busybox config to one based on defconfig and adjust inittab
Use the busybox defconfig as the base for our busybox build as it
provides a number of useful features over the stripped-down busybox
config that was used before.  Contrary to the previously used config
the defconfig installs a number of applets including ntpd to /usr, so
that requires adjusting the inittab.

Author: Karsten Merker <merker@debian.org>